Fixes #8: oc-inject: add --java support
Tweaks to load additional dependencies for JDK tools used to
troubleshoot a Java container. Enabled if --java version is provided
or when the executable requests libjli.so.

Tested with jps, jstat, jstack.

Turns out that Java programs do not like to be loaded via ld-linux-x86-64.so.2.
This forces us to use ld-linux-x86-64.so.2 from the container (not injected)
and correspondingly also the version of glibc from the container.

Since the container presumably has a Java program and therefore libc,
and the JDK tool version must closely match the container JRE version
anyways, this is not as serious a problem as it sounds.